Adding Channel Multiplexer
Added a function that acts as a Go Channel multiplexer. This is because
I end up doing this alot in some of the userspace tools when I need to
capture one message and do two different things with it each in their
own goroutine (Save to database AND send on another interface for example).